# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: 商用地理信息虚假内容检查与反投喂平台 API
  description: |
    ## 概述
    本API提供地理信息虚假内容检测服务，支持单条/批量检测、规则管理、审核标注、统计报表等功能。
    
    ## 认证方式
    - **JWT Token**: 用于Web端用户认证，在Header中添加 `Authorization: Bearer <token>`
    - **API Key**: 用于服务端对接，在Header中添加 `X-API-Key: <api_key>`
    
    ## 请求限制
    - 基础版: 10,000次/天
    - 专业版: 50,000次/天
    - 企业版: 不限制
    
    ## 错误码
    | 错误码 | 说明 |
    |--------|------|
    | 400 | 请求参数错误 |
    | 401 | 未授权/Token过期 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 429 | 请求频率超限 |
    | 500 | 服务器内部错误 |
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@geo-fake-detection.com
servers:
  - url: https://api.geo-fake-detection.com/v1
    description: 生产环境
  - url: https://api-staging.geo-fake-detection.com/v1
    description: 测试环境
...
paths:
...
  /rules:
    get:
      tags: [Rules]
      summary: 规则列表
      security:
        - bearerAuth: []
      parameters:
        - name: rule_type
          in: query
          schema:
            type: string
            enum: [geo, text, simhash, semantic]
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RuleConfig'

    post:
      tags: [Rules]
      summary: 创建规则
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleConfigCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'
  /rules/{id}:
    get:
      tags: [Rules]
      summary: 规则详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'

    put:
      tags: [Rules]
      summary: 更新规则
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleConfigUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfig'

    delete:
      tags: [Rules]
      summary: 删除规则
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /rules/test:
    post:
      tags: [Rules]
      summary: 测试规则
      description: 使用测试数据验证规则效果
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rule, test_data]
              properties:
                rule:
                  $ref: '#/components/schemas/RuleConfigCreate'
                test_data:
                  type: array
                  items:
                    $ref: '#/components/schemas/DetectionRequest'
      responses:
        '200':
          description: 测试结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  hit_count:
                    type: integer
                  total_count:
                    type: integer
                  hit_rate:
                    type: number
                  examples:
                    type: array
                    items:
                      type: object

  # ============================================
  # 关键词管理
  # ============================================
...
  /alerts/rules:
    get:
      tags: [Alerts]
      summary: 告警规则列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertRule'

    post:
      tags: [Alerts]
      summary: 创建告警规则
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
  /alerts/rules/{id}:
    put:
      tags: [Alerts]
      summary: 更新告警规则
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    DetectionRequest:
      type: object
      required: [device_id, user_id, timestamp, geo, content]
      properties:
        device_id:
          type: string
          description: 设备唯一标识
        user_id:
          type: string
          description: 用户唯一标识
        session_id:
          type: string
          description: 会话ID
        timestamp:
          type: integer
          format: int64
          description: Unix时间戳（秒）
        geo:
          type: object
          required: [latitude, longitude]
          properties:
            latitude:
              type: number
              minimum: -90
              maximum: 90
            longitude:
              type: number
              minimum: -180
              maximum: 180
            accuracy:
              type: number
              description: 精度（米）
            source:
              type: string
              enum: [GPS, WIFI, CELL, IP]
        content:
          type: object
          required: [text]
          properties:
            text:
              type: string
              maxLength: 5000
            type:
              type: string
              enum: [review, checkin, poi, comment]
              default: review
        metadata:
          type: object
          properties:
            ip:
              type: string
            app_version:
              type: string
            user_agent:
              type: string
            platform:
              type: string
...
    RuleConfig:
      type: object
      properties:
        id:
          type: integer
        rule_type:
          type: string
          enum: [geo, text, simhash, semantic]
        rule_name:
          type: string
        rule_code:
          type: string
        rule_config:
          type: object
        description:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        hit_count:
          type: integer
        created_at:
          type: string
          format: date-time
    RuleConfigCreate:
      type: object
      required: [rule_type, rule_name, rule_config]
      properties:
        rule_type:
          type: string
          enum: [geo, text, simhash, semantic]
        rule_name:
          type: string
        rule_code:
          type: string
        rule_config:
          type: object
        description:
          type: string
        enabled:
          type: boolean
          default: true
        priority:
          type: integer
          default: 100
    RuleConfigUpdate:
      type: object
      properties:
        rule_name:
          type: string
        rule_config:
          type: object
        description:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
...
    AlertRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        alert_type:
          type: string
        condition_config:
          type: object
        notification_config:
          type: object
        severity:
          type: string
        enabled:
          type: boolean
        last_triggered_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    AlertRuleCreate:
      type: object
      required: [name, alert_type, condition_config, notification_config]
      properties:
        name:
          type: string
        description:
          type: string
        alert_type:
          type: string
          enum: [threshold, anomaly, trend]
        condition_config:
          type: object
          description: 告警条件配置
          example:
            metric: "fake_rate"
            operator: ">"
            threshold: 0.3
            duration: "5m"
        notification_config:
          type: object
          description: 通知配置
          example:
            channels: ["email", "webhook"]
            emails: ["admin@example.com"]
            webhook_url: "https://hooks.example.com/alert"
        severity:
          type: string
          enum: [info, warning, critical]
          default: warning
        enabled:
          type: boolean
          default: true
        cooldown_minutes:
          type: integer
          default: 60
    AlertRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        condition_config:
          type: object
        notification_config:
          type: object
        severity:
          type: string
        enabled:
          type: boolean
        cooldown_minutes:
          type: integer